refactor: pages 컴포넌트 안으로 이동
